Change create group to dialog
